Packages names renamed
* lowercase for all names
* repository package created
