+++ backend/src/main/java/student/aschm22s/hbrsiCalGenerator/hbrsiCalGenerator/models/DAOObjects/VeranstaltungsIdsAndEmail.java
@@ -1,4 +1,4 @@
-package student.aschm22s.hbrsiCalGenerator.hbrsiCalGenerator.Models.DAOObjects;
+package student.aschm22s.hbrsiCalGenerator.hbrsiCalGenerator.models.DAOObjects;
 
 import java.util.List;
