+++ backend/src/main/java/student/aschm22s/hbrsiCalGenerator/hbrsiCalGenerator/repository/AbusingIPAddressRepo.java
@@ -1,7 +1,7 @@
-package student.aschm22s.hbrsiCalGenerator.hbrsiCalGenerator.DBRepo;
+package student.aschm22s.hbrsiCalGenerator.hbrsiCalGenerator.repository;
 
 import org.springframework.data.repository.CrudRepository;
-import student.aschm22s.hbrsiCalGenerator.hbrsiCalGenerator.Models.AbusingIPAddress;
+import student.aschm22s.hbrsiCalGenerator.hbrsiCalGenerator.models.AbusingIPAddress;
 
 public interface AbusingIPAddressRepo extends CrudRepository<AbusingIPAddress, Integer> {
     int countAbusingIPAddressByIpAddress(String ip);
