+++ backend/src/main/java/student/aschm22s/hbrsiCalGenerator/hbrsiCalGenerator/repository/StundenplanDateMNRepo.java
@@ -1,8 +1,8 @@
-package student.aschm22s.hbrsiCalGenerator.hbrsiCalGenerator.DBRepo;
+package student.aschm22s.hbrsiCalGenerator.hbrsiCalGenerator.repository;
 
 import org.springframework.data.repository.CrudRepository;
-import student.aschm22s.hbrsiCalGenerator.hbrsiCalGenerator.Models.StundenplanDatumMN;
-import student.aschm22s.hbrsiCalGenerator.hbrsiCalGenerator.Models.StundenplanEintrag;
+import student.aschm22s.hbrsiCalGenerator.hbrsiCalGenerator.models.StundenplanDatumMN;
+import student.aschm22s.hbrsiCalGenerator.hbrsiCalGenerator.models.StundenplanEintrag;
 
 public interface StundenplanDateMNRepo extends CrudRepository<StundenplanDatumMN, Integer> {
     Iterable<StundenplanDatumMN> findByStundenplanEintrag(StundenplanEintrag stundenplanEintrag);
