+++ backend/src/main/java/student/aschm22s/hbrsiCalGenerator/hbrsiCalGenerator/repository/VeranstaltungsRepo.java
@@ -1,8 +1,8 @@
-package student.aschm22s.hbrsiCalGenerator.hbrsiCalGenerator.DBRepo;
+package student.aschm22s.hbrsiCalGenerator.hbrsiCalGenerator.repository;
 
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
-import student.aschm22s.hbrsiCalGenerator.hbrsiCalGenerator.Models.Veranstaltung;
+import student.aschm22s.hbrsiCalGenerator.hbrsiCalGenerator.models.Veranstaltung;
 
 @Repository
 public interface VeranstaltungsRepo extends CrudRepository<Veranstaltung, Integer> {
